// src/model/formOut.ts
/**
 * 输出表单定义
 */
export interface IFormOut {
  /**
   * 接口出入参模型声明代码
   */
  decCode: string;
  /**
   * 接口定义样例代码
   */
  defCode: string;
  /**
   * 接口调用样例代码
   */
  useCode: string;
  /**
   * 导入具体模型代码
   */
  importCode: string;
}